+++ HISTORY.rst
@@ -17,6 +17,7 @@ Unreleased
 * ALCN-1189/ALCYN-338: Add fax service and send sale order confirmation through it, if sale channel is fax.
 * ALCYN-184: Add `web` to sale_channel. Remove sale_channel_invisible.
 * ALCYN-449: Set a default bin_checksum, set the bin_checksum_2 and update parking locators
+* ALCYN-448: Fix duplicating sequences in sale_order lines on additional products
 
 **Bugfixes**
 
@@ -55,8 +56,6 @@ Unreleased
 * ALCN-1226: Push product on sales order in backorder
 * ALCYN-436: Improve web service new sale order from Magento, splitting job using confirm in background job.
 * ALCN-757: Accounting cut-off
-
-* ALCYN-448: Fix duplicating sequences in sale_order lines on additional products
 
 **Bugfixes**